ENH: Toggle between pattern alternatives when executed again
As we cannot easily reverse the transformation, store the two alternatives in script-local variables and use those for recall. (That means that the recall only works for the very last transformation, but that should be fine. If the pattern can be found in the search history, you could just recall it and trigger the toggling again.)