CHG: Switch <C-G>y to <C-G>Y and add <C-G>y{x} variant that allows to pass the register to yank the command-line to
Add <silent> to both mapping definitions; this doesn't matter for the original <C-g>Y, but gets rid of the <Plug>-definition in the command-line while waiting for the register. Fortunately, this mapping does not modify the command-line itself, so the missed update of it caused by <silent> doesn't matter here.