Add the :s_e flag to the (last) :substitute (or similar) command
Usually, one does not add the /e flag; an error in a single buffer is easy to ignore and does provide some feedback - maybe the pattern wasn't right. But when iterating multiple buffers, the accumulated errors are more distracting, and also much more likely to happen, as the granularity of the various iteration commands is very coarse, and there's a high likelihood that some unrelated buffer is in there. Therefore, the DWIM action is automatically appending the /e flag for the user; if substitution over all buffers really is crucial, it's easy to remove (<End><BS>).
If the command-line consists of a pipeline of substitutions, the /e flag is only applied to the last, as the error on a previous substitution does have the implication of aborting the pipeline. If one wanted a best-effort pipeline application of multiple substitutions, the /e flag would have been added already.