ENH: Also add the search pattern toggled via normal mode <A-/>, <A-?>, <A-(> to the search history
For the command-line mode mappings, this is a natural side effect of editing the command-line. But for the normal mode mappings, the result is only in the @/ register, so used for searches, but overwritten by any future search, and it cannot be found in the history. Better be consistent and allow easy recall by explicitly putting the toggle result into the search history, too.